refactor: add manipulation and apply it
